src/Health_AI_Agent/requirements.txt
fastapi
uvicorn
openai
psycopg2-binary
google-cloud-bigquery
langchain
tiktoken
pydantic
python-dotenv
pgvector
